move error checker methods to error_checker.go
