+++ common/persistence/sql/sqlplugin/sqlite/db.go
@@ -85,27 +85,3 @@ func (mdb *db) MaxAllowedTTL() (*time.Duration, error) {
 func (mdb *db) SupportsAsyncTransaction() bool {
 	return false
 }
-
-// IsDupEntryError verify if the error is a duplicate entry error
-func (mdb *db) IsDupEntryError(err error) bool {
-	//TODO implement me
-	panic("implement me")
-}
-
-// IsNotFoundError verify if the error is a not found error
-func (mdb *db) IsNotFoundError(err error) bool {
-	//TODO implement me
-	panic("implement me")
-}
-
-// IsTimeoutError verify if the error is a timeout error
-func (mdb *db) IsTimeoutError(err error) bool {
-	//TODO implement me
-	panic("implement me")
-}
-
-// IsThrottlingError verify if the error is a throttling error
-func (mdb *db) IsThrottlingError(err error) bool {
-	//TODO implement me
-	panic("implement me")
-}
