+++ common/persistence/sql/sqlplugin/sqlite/error_checker.go
@@ -1,0 +1,25 @@
+package sqlite
+
+// IsDupEntryError verify if the error is a duplicate entry error
+func (mdb *db) IsDupEntryError(err error) bool {
+	//TODO implement me
+	panic("implement me")
+}
+
+// IsNotFoundError verify if the error is a not found error
+func (mdb *db) IsNotFoundError(err error) bool {
+	//TODO implement me
+	panic("implement me")
+}
+
+// IsTimeoutError verify if the error is a timeout error
+func (mdb *db) IsTimeoutError(err error) bool {
+	//TODO implement me
+	panic("implement me")
+}
+
+// IsThrottlingError verify if the error is a throttling error
+func (mdb *db) IsThrottlingError(err error) bool {
+	//TODO implement me
+	panic("implement me")
+}
